+++ trace/driver.go
@@ -55,27 +55,6 @@ type (
 
 		// Credentials events
 		OnGetCredentials func(DriverGetCredentialsStartInfo) func(DriverGetCredentialsDoneInfo)
-
-		// Deprecated: has no effect now
-		OnConnUsagesChange func(DriverConnUsagesChangeInfo)
-		// Deprecated: has no effect now
-		OnConnStreamUsagesChange func(DriverConnStreamUsagesChangeInfo)
-		// Deprecated: has no effect now
-		OnConnRelease func(DriverConnReleaseStartInfo) func(DriverConnReleaseDoneInfo)
-		// Deprecated: has no effect now
-		OnClusterInit func(DriverClusterInitStartInfo) func(DriverClusterInitDoneInfo)
-		// Deprecated: has no effect now
-		OnClusterClose func(DriverClusterCloseStartInfo) func(DriverClusterCloseDoneInfo)
-		// Deprecated: has no effect now
-		OnClusterGet func(DriverClusterGetStartInfo) func(DriverClusterGetDoneInfo)
-		// Deprecated: has no effect now
-		OnClusterInsert func(DriverClusterInsertStartInfo) func(DriverClusterInsertDoneInfo)
-		// Deprecated: has no effect now
-		OnClusterRemove func(DriverClusterRemoveStartInfo) func(DriverClusterRemoveDoneInfo)
-		// Deprecated: has no effect now
-		OnPessimizeNode func(DriverPessimizeNodeStartInfo) func(DriverPessimizeNodeDoneInfo)
-		// Deprecated: has no effect now
-		OnUnpessimizeNode func(DriverUnpessimizeNodeStartInfo) func(DriverUnpessimizeNodeDoneInfo)
 	}
 )
 
@@ -129,34 +108,6 @@ type EndpointInfo interface {
 }
 
 type (
-	// Deprecated: has no effect now
-	DriverClusterInsertStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context  *context.Context
-		Endpoint EndpointInfo
-	}
-	// Deprecated: has no effect now
-	DriverClusterInsertDoneInfo struct {
-		Inserted bool
-		State    ConnState
-	}
-	// Deprecated: has no effect now
-	DriverClusterRemoveStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context  *context.Context
-		Endpoint EndpointInfo
-	}
-	// Deprecated: has no effect now
-	DriverClusterRemoveDoneInfo struct {
-		Removed bool
-		State   ConnState
-	}
 	DriverConnStateChangeStartInfo struct {
 		Endpoint EndpointInfo
 		State    ConnState
@@ -275,29 +226,6 @@ type (
 	DriverConnAllowDoneInfo struct {
 		State ConnState
 	}
-	// Deprecated: has no effect now
-	DriverConnReleaseStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context  *context.Context
-		Endpoint EndpointInfo
-	}
-	// Deprecated: has no effect now
-	DriverConnReleaseDoneInfo struct {
-		Error error
-	}
-	// Deprecated: has no effect now
-	DriverConnUsagesChangeInfo struct {
-		Endpoint EndpointInfo
-		Usages   int
-	}
-	// Deprecated: has no effect now
-	DriverConnStreamUsagesChangeInfo struct {
-		Endpoint EndpointInfo
-		Usages   int
-	}
 	DriverConnInvokeStartInfo struct {
 		// Context make available context in trace callback function.
 		// Pointer to context provide replacement of context in trace callback function.
@@ -329,43 +257,6 @@ type (
 		State ConnState
 		Error error
 	}
-	// Deprecated: has no effect now
-	DriverClusterInitStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context *context.Context
-	}
-	// Deprecated: has no effect now
-	DriverClusterInitDoneInfo struct {
-		Error error
-	}
-	// Deprecated: has no effect now
-	DriverClusterCloseStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context *context.Context
-	}
-	// Deprecated: has no effect now
-	DriverClusterCloseDoneInfo struct {
-		Error error
-	}
-	// Deprecated: has no effect now
-	DriverClusterGetStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context *context.Context
-	}
-	// Deprecated: has no effect now
-	DriverClusterGetDoneInfo struct {
-		Endpoint EndpointInfo
-		Error    error
-	}
 	DriverBalancerInitStartInfo struct {
 		// Context make available context in trace callback function.
 		// Pointer to context provide replacement of context in trace callback function.
@@ -396,35 +287,6 @@ type (
 	DriverBalancerChooseEndpointDoneInfo struct {
 		Endpoint EndpointInfo
 		Error    error
-	}
-	// Deprecated: has no effect now
-	DriverPessimizeNodeStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context  *context.Context
-		Endpoint EndpointInfo
-		State    ConnState
-		Cause    error
-	}
-	// Deprecated: has no effect now
-	DriverPessimizeNodeDoneInfo struct {
-		State ConnState
-	}
-	// Deprecated: has no effect now
-	DriverUnpessimizeNodeStartInfo struct {
-		// Context make available context in trace callback function.
-		// Pointer to context provide replacement of context in trace callback function.
-		// Warning: concurrent access to pointer on client side must be excluded.
-		// Safe replacement of context are provided only inside callback function
-		Context  *context.Context
-		Endpoint EndpointInfo
-		State    ConnState
-	}
-	// Deprecated: has no effect now
-	DriverUnpessimizeNodeDoneInfo struct {
-		State ConnState
 	}
 	DriverRepeaterWakeUpStartInfo struct {
 		// Context make available context in trace callback function.